Query: Make GetMappedProjectionTypes to return based on actual projections
Resolves #10035

Issue here was, our method to calculated projection types was not entirely accurate. It was not robust enough to handle a case where there is project star & projection both at the same time (Sql Generation handles that pretty well).
Due to the bug in the method, when we encouter project star with projection list, we returned based on projection only ignoring the star table. At present this kind of structure only arises with RowNumberPaging which ads RowNumber Expression on projection with star.
Hence we got incorrect type in first slot causing InvalidCastException.

Also changed type for RowNumberExpression since, RowNumber is of type bigint in SqlServer.

This PR brings projection & materialization on same page w.r.t. CLR types.

All tests passed when using `TypedRelationalValueBufferFactoryFactory` in SqlServer provider.

+++ src/EFCore.SqlServer/Query/Expressions/Internal/RowNumberExpression.cs
@@ -47,7 +47,7 @@ namespace Microsoft.EntityFrameworkCore.Query.Expressions.Internal
         ///     This API supports the Entity Framework Core infrastructure and is not intended to be used
         ///     directly from your code. This API may change or be removed in future releases.
         /// </summary>
-        public override Type Type => typeof(int);
+        public override Type Type => typeof(long);
 
         /// <summary>
         ///     This API supports the Entity Framework Core infrastructure and is not intended to be used
